Implement WebGL console command support
