+++ CUDLR/Scripts/WebGL_Relay.cs
@@ -1,0 +1,54 @@
+﻿using AOT;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.InteropServices;
+using System.Text;
+using UnityEngine;
+
+namespace CUDLR
+{
+    /// <summary>
+    /// WebGLとCUDLR.Consoleを中継する
+    /// </summary>
+    public static class WebGL_Relay
+    {
+        /// <summary>
+        /// 初期化。ブラウザ側 javascript にコールバック設定
+        /// </summary>
+        public static void Initialize()
+        {
+#if UNITY_WEBGL && !UNITY_EDITOR
+            Debug.LogFormat("Initialize CUDLR WebGL Relay");
+            SetCudlrCallbackAndCreateInput(RunCommand);
+            Console.Update();
+#endif
+        }
+
+        public static void Log(string text)
+        {
+#if UNITY_WEBGL && !UNITY_EDITOR
+            ShowLog(text);
+#endif
+        }
+
+#if UNITY_WEBGL && !UNITY_EDITOR
+        /// <summary>
+        /// 汎用コマンド実行コールバック
+        /// </summary>
+        /// <param name="commandline"></param>
+        [MonoPInvokeCallback(typeof(Func<string,string>))]
+        static private string RunCommand(string commandline)
+        {
+            Console.Run(commandline);
+            return Console.GetResult();
+        }
+
+        [DllImport("__Internal")]
+        static extern void SetCudlrCallbackAndCreateInput(Func<string,string> callback);
+
+        [DllImport("__Internal")]
+        static extern string ShowLog(string text);
+#endif
+    }
+}
